normata tranne i 25

+++ 02_Printf/out_char.c
@@ -6,31 +6,29 @@
 /*   By: jfranco <jfranco@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 13:22:44 by jfranco           #+#    #+#             */
-/*   Updated: 2024/11/08 17:23:22 by jfranco          ###   ########.fr       */
+/*   Updated: 2024/11/08 17:51:39 by jfranco          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int out_char(int c)
+int	out_char(int c)
 {
 	return (write(1, &c, 1));
 }
 
-int print_ad(void *ptr)
+int	print_ad(void *ptr)
 {
-    int count;
-    uintptr_t adress;
-   
-   count = 0;
-  if (ptr == NULL)
-  {
-	  return (write(1, "0x0", 3));
-  } 
-   adress = (uintptr_t)ptr;
-   count += write(1, "0x", 2);
-   count += out_hex(adress);
-   return (count);
+	int long	count;
+	uintptr_t	adress;
+
+	count = 0;
+	if (ptr == NULL)
+	{
+		return (write(1, "0x0", 3));
+	}
+	adress = (uintptr_t)ptr;
+	count += write(1, "0x", 2);
+	count += out_hex(adress);
+	return (count);
 }
-
-
